+++ src/domain/value/space_image_id.go
@@ -1,5 +1,5 @@
 package value
 
 type SpaceImageId struct {
-	Value uint
+	Value uint64
 }
